requirements.txt
matplotlib==3.3.2
streamlit==0.78.0
pandas==1.2.0
numpy==1.19.2
Pillow==8.2.0
scikit_learn==0.24.1
